"dataValues" can be greater as 127, but then they need to be passed as durability (which is a short)

+++ src/main/java/com/geNAZt/RegionShop/Config/Sub/Item.java
@@ -7,7 +7,7 @@ package com.geNAZt.RegionShop.Config.Sub;
  */
 public class Item {
     public Integer itemID;
-    public Byte dataValue;
+    public Short dataValue;
     public Float buy;
     public Float sell;
     public Integer maxItemRecalc;
